requirements.txt
python_decouple==3.1
boto3==1.4.7
dronekit==2.9.1
